requirements.txt
fastapi==0.111.0
uvicorn==0.29.0

google-generativeai==0.3.2

langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10

sentence-transformers==2.2.2
chromadb==0.4.22

python-dotenv==1.0.0
